refactor: change the name of load function files
