+++ src/op/load/mod.rs
@@ -1,7 +1,7 @@
-mod load_op;
-mod load_rg;
+mod load_ld;
+mod load_lrg;
 mod load_lea;
 
-pub use load_op::load;
-pub use load_rg::load_register;
+pub use load_ld::load;
+pub use load_lrg::load_register;
 pub use load_lea::load_effective_address;
